common_game::components::planet

Trait PlanetAI

pub trait PlanetAI: Send {
    // Required methods
    fn handle_orchestrator_msg(
        &mut self,
        state: &mut PlanetState,
        generator: &Generator,
        combinator: &Combinator,
        msg: OrchestratorToPlanet,
    ) -> Option<PlanetToOrchestrator>;
    fn handle_explorer_msg(
        &mut self,
        state: &mut PlanetState,
        generator: &Generator,
        combinator: &Combinator,
        msg: ExplorerToPlanet,
    ) -> Option<PlanetToExplorer>;
    fn handle_asteroid(
        &mut self,
        state: &mut PlanetState,
        generator: &Generator,
        combinator: &Combinator,
    ) -> Option<Rocket>;
    fn start(&mut self, state: &PlanetState);
    fn stop(&mut self, state: &PlanetState);
}

The trait that defines the behaviour of a planet.

Structs implementing this trait are intended to be passed to the Planet constructor, so that the handlers can be invoked by the planet internal logic when certain messages are received on any of the planet channels.

The handlers can alter the planet state by accessing the state parameter, which is passed to the methods as a mutable borrow. A response can be sent by returning an optional message of the correct type, that will be forwarded to the associated channel passed on planet construction.

Required Methods

fn handle_orchestrator_msg( &mut self, state: &mut PlanetState, generator: &Generator, combinator: &Combinator, msg: OrchestratorToPlanet, ) -> Option<PlanetToOrchestrator>

Handler for messages received by the orchestrator (receiving end of the OrchestratorToPlanet channel). The following messages will not invoke this handler:

OrchestratorToPlanet::StartPlanetAI (see PlanetAI::start)

OrchestratorToPlanet::StopPlanetAI (see PlanetAI::stop)

OrchestratorToPlanet::Asteroid (see PlanetAI::handle_asteroid)

OrchestratorToPlanet::IncomingExplorerRequest, as this will be handled automatically by the planet

OrchestratorToPlanet::OutgoingExplorerRequest (same as previous one)

Check PlanetAI docs for general meaning of the parameters and return type.

fn handle_explorer_msg( &mut self, state: &mut PlanetState, generator: &Generator, combinator: &Combinator, msg: ExplorerToPlanet, ) -> Option<PlanetToExplorer>

Handler for all messages received by an explorer (receiving end of the ExplorerToPlanet channel).

Check PlanetAI docs for general meaning of the parameters and return type.

fn handle_asteroid( &mut self, state: &mut PlanetState, generator: &Generator, combinator: &Combinator, ) -> Option<Rocket>

This handler will be invoked when a OrchestratorToPlanet::Asteroid message is received. It’s important to handle Asteroid messages correctly, as this will the determine the planet survival.

Returns

In order to survive, an owned Rocket must be returned from this method; if None is returned instead, the planet will (or should) be destroyed by the orchestrator

fn start(&mut self, state: &PlanetState)

This method will be invoked when a OrchestratorToPlanet::StartPlanetAI is received, but only if the planet is currently in a stopped state.

Start messages received when planet is already running are ignored.

fn stop(&mut self, state: &PlanetState)

This method will be invoked when a OrchestratorToPlanet::StopPlanetAI is received, but only if the planet is currently in a running state.

Stop messages received when planet is already stopped are ignored.