common_game::components

Module resource

Resource Module

Defines all basic and complex resources.

Every resource has a struct and implements the resource trait.

The primary public items for other modules are the Resource trait, the enums that aggregate the generated types, such as BasicResource, ComplexResource and GenericResource for passing the actual Resource around and BasicResourceType,ComplexResourceType and ResourceType for passing the names and acts like phantoms of a resource. There exist two structs Generator and Combinator which are given to the planet to generate and combine resources, these structs contain the signatures of the available recipes, these recipes are only updated in the constructor of the planet To generate of combinator resources use the available methods in either structs, these return a Result There exist an enum to pass around the request to the planet to generate a complex resource ComplexResourceRequest

Structs

AIPartner

Carbon

Combinator

contains all the recipes available to a planet and enables the creation of complex resources

Diamond

Dolphin

Generator

contains all the recipes available to a planet and enables the creation of basic resources

Hydrogen

Life

Oxygen

Robot

Silicon

Water

Enums

BasicResource

Gives the choice between every possible basic resource

BasicResourceType

Identifies a BasicResource without actually containing the underlying resource,

ComplexResource

Gives the choice between every possible complex resource

ComplexResourceRequest

Gives a structured way to pass around the request to produce a complex resource

ComplexResourceType

Identifies a ComplexResource without actually containing the underlying resource,

GenericResource

Contains a resource which could be both BasicResource and ComplexResource

ResourceType

Identifies a resource which could be both BasicResourceType and ComplexResourceType without actually containing the underlying resource,

Traits

Resource

Gives the necessary methods to print out the resource