common_game::components::resource

Enum BasicResource

pub enum BasicResource {
    Oxygen(Oxygen),
    Hydrogen(Hydrogen),
    Carbon(Carbon),
    Silicon(Silicon),
}

An enum that provides a unified type for all possible basic resources.

This enum wraps every generated basic resource struct (e.g., Oxygen, Hydrogen) into a single type. It is useful when you need to store or pass around any basic resource without knowing its specific concrete type at compile time.

Variants

Oxygen(Oxygen)

Hydrogen(Hydrogen)

Carbon(Carbon)

Silicon(Silicon)

Implementations

impl BasicResource

pub fn get_type(&self) -> BasicResourceType

Returns the BasicResourceType of this resource.

pub fn to_oxygen(self) -> Result<Oxygen, String>

Attempts to convert the BasicResource into a $basic.

Returns

Ok($basic) if the resource is $basic.

Errors

Returns an error if the resource is of a different type.

pub fn to_hydrogen(self) -> Result<Hydrogen, String>

Attempts to convert the BasicResource into a $basic.

Returns

Ok($basic) if the resource is $basic.

Errors

Returns an error if the resource is of a different type.

pub fn to_carbon(self) -> Result<Carbon, String>

Attempts to convert the BasicResource into a $basic.

Returns

Ok($basic) if the resource is $basic.

Errors

Returns an error if the resource is of a different type.

pub fn to_silicon(self) -> Result<Silicon, String>

Attempts to convert the BasicResource into a $basic.

Returns

Ok($basic) if the resource is $basic.

Errors

Returns an error if the resource is of a different type.