common_game::components::resource

Trait Bag

pub trait Bag<T> {
    // Required methods
    fn is_empty(&self) -> bool;
    fn contains_complex(&self, complex: ComplexResourceType) -> bool;
    fn contains_basic(&self, basic: BasicResourceType) -> bool;
    fn contains(&self, generic: ResourceType) -> bool;
    fn size(&self) -> u32;
    fn capacity(&self) -> u32;
    fn get_count(&self, generic: ResourceType) -> u32;
    fn get_count_complex(&self, complex: ComplexResourceType) -> u32;
    fn get_count_basic(&self, basic: BasicResourceType) -> u32;
    fn get_content(&self) -> T;
}

Required Methods

fn is_empty(&self) -> bool

fn contains_complex(&self, complex: ComplexResourceType) -> bool

fn contains_basic(&self, basic: BasicResourceType) -> bool

fn contains(&self, generic: ResourceType) -> bool

fn size(&self) -> u32

fn capacity(&self) -> u32

fn get_count(&self, generic: ResourceType) -> u32

fn get_count_complex(&self, complex: ComplexResourceType) -> u32

fn get_count_basic(&self, basic: BasicResourceType) -> u32

fn get_content(&self) -> T

this method serves the purpose to the expose the internal structure of the bag and is implementation dependent,