common_game

Module protocols

Protocol for actors to communicate.

Includes a definition for all messages and documentation on how to use them. Graphs visualizing the expected flow of messages can be viewed on GitHub

Modules

orchestrator_explorer

Orchestrator and Explorer protocol messages

orchestrator_planet

Orchestrator and Planet protocol messages

planet_explorer

Planet and Explorer protocol messages