common_game::protocols::messages

Enum ExplorerToOrchestrator

pub enum ExplorerToOrchestrator<T> {
    StartExplorerAIResult {
        explorer_id: u32,
    },
    KillExplorerAIResult {
        explorer_id: u32,
    },
    ResetExplorerAIResult {
        explorer_id: u32,
    },
    MovedToPlanetResult {
        explorer_id: u32,
    },
    CurrentPlanetResult {
        explorer_id: u32,
        planet_id: u32,
    },
    SupportedResourceResult {
        explorer_id: u32,
        supported_resources: HashSet<BasicResourceType>,
    },
    SupportedCombinationResult {
        explorer_id: u32,
        combination_list: HashSet<ComplexResourceType>,
    },
    GenerateResourceResponse {
        explorer_id: u32,
        generated: Result<(), ()>,
    },
    CombineResourceResponse {
        explorer_id: u32,
        generated: Result<(), ()>,
    },
    BagContentResponse {
        explorer_id: u32,
        bag_content: T,
    },
    NeighborsRequest {
        explorer_id: u32,
        current_planet_id: u32,
    },
    TravelToPlanetRequest {
        explorer_id: u32,
        current_planet_id: u32,
        dst_planet_id: u32,
    },
}

Messages sent by an Explorer to the Orchestrator.

Variants

StartExplorerAIResult

Acknowledge of OrchestratorToExplorer::StartExplorerAI

Fields

explorer_id: u32

KillExplorerAIResult

Acknowledge of OrchestratorToExplorer::KillExplorerAI

Fields

explorer_id: u32

ResetExplorerAIResult

Acknowledge of OrchestratorToExplorer::ResetExplorerAI

Fields

explorer_id: u32

MovedToPlanetResult

Acknowledge of OrchestratorToExplorer::MoveToPlanet

Fields

explorer_id: u32

CurrentPlanetResult

This variant is used to send the ID of the current planet on which the Explorer is located

Fields

explorer_id: u32

planet_id: u32

SupportedResourceResult

This variant is used to send the list of the available BasicResourceType in the planet

Fields

explorer_id: u32

supported_resources: HashSet<BasicResourceType>

SupportedCombinationResult

This variant is used to send the list of the available ComplexResourceType in the planet

Fields

explorer_id: u32

combination_list: HashSet<ComplexResourceType>

GenerateResourceResponse

This variant holds a Result for the Orchestrator to know if the requested BasicResource has been generated

Fields

explorer_id: u32

generated: Result<(), ()>

CombineResourceResponse

This variant holds a Result for the Orchestrator to know if the requested ComplexResource has been generated

Fields

explorer_id: u32

generated: Result<(), ()>

BagContentResponse

This message is for passing around the bag content and has been implemented with a generic type to let the group the freedom to implement the methods on it

Example

use std::collections::HashMap;
use common_game::components::resource::{ComplexResourceType, BasicResourceType};
use common_game::protocols::messages::ExplorerToOrchestrator;

pub struct DummyBag {
    pub complex: HashMap<ComplexResourceType, u32>,
    pub basic: HashMap<BasicResourceType, u32>,
}

let message = ExplorerToOrchestrator::BagContentResponse {
    explorer_id: 1,
    bag_content: DummyBag {
        complex: HashMap::new(),
        basic: HashMap::new(),
    }
};

Fields

explorer_id: u32

bag_content: T

NeighborsRequest

This variant asks the Orchestrator for the list of neighbors Planets to travel to

Fields

explorer_id: u32

current_planet_id: u32

TravelToPlanetRequest

This variant asks the Orchestrator to be sent to the specified Planet

Fields

explorer_id: u32

current_planet_id: u32

dst_planet_id: u32

Implementations

impl<T> ExplorerToOrchestrator<T>

pub fn explorer_id(&self) -> u32

Helper method to extract the explorer_id field from any message variant without needing to match a specific one.