common_game::protocols::messages

Enum OrchestratorToExplorer

pub enum OrchestratorToExplorer {
    StartExplorerAI,
    ResetExplorerAI,
    KillExplorerAI,
    MoveToPlanet {
        sender_to_new_planet: Option<Sender<ExplorerToPlanet>>,
    },
    CurrentPlanetRequest,
    SupportedResourceRequest,
    SupportedCombinationRequest,
    GenerateResourceRequest {
        to_generate: BasicResourceType,
    },
    CombineResourceRequest(ComplexResourceRequest),
    BagContentRequest,
    NeighborsResponse {
        neighbors: Vec<u32>,
    },
}

Messages sent by the Orchestrator to an Explorer.

Variants

StartExplorerAI

This variant is used to start the Explorer AI

ResetExplorerAI

This variant is used to reset the Explorer AI

KillExplorerAI

This variant is used to kill the Explorer AI

MoveToPlanet

This variant is used to send a Sender to the new planet

Fields

sender_to_new_planet: Option<Sender<ExplorerToPlanet>>

CurrentPlanetRequest

This variant is used to ask the ID of the Planet in which the Explorer is currently located

SupportedResourceRequest

This variant is used to enforce the Explorer to ask the supported Resources on the Planet

SupportedCombinationRequest

This variant is used to enforce the Explorer to ask the supported Combinations on the Planet

GenerateResourceRequest

This variant is used to enforce the Explorer to ask the Planet to Generate a BasicResource

Fields

to_generate: BasicResourceType

CombineResourceRequest(ComplexResourceRequest)

This variant is used to enforce the Explorer to ask the Planet to Generate a ComplexResource using the ComplexResourceRequest

BagContentRequest

This variant is used to ask the content of the Explorer Bag

NeighborsResponse

This variant is used to send to the Explorer its neighbors’ IDs

Fields

neighbors: Vec<u32>