common_game::protocols

Module orchestrator_explorer

Orchestrator and Explorer protocol messages

Defines the types of messages exchanged of the full duplex communication channel between the Orchestrator and the Explorers For a more detailed view of the interactions between these two entities, visit the communications diagrams

Enums

ExplorerToOrchestrator

This enum describes all possible messages from an Explorer to the Orchestrator

ExplorerToOrchestratorKind

OrchestratorToExplorer

This enum describes all possible messages from the Orchestrator to an Explorer

OrchestratorToExplorerKind